requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=2.1.6
langchain-core>=0.1.0
langchain-community>=0.0.20
google-generativeai>=0.8.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
plotly>=5.15.0
